moving BookForm to container dir
